[refactor] s3와 관련없는 함수 분리
