+++ libs/nest-modules-goods/src/goods/goods-common-info.service.ts
@@ -8,8 +8,6 @@ import { GoodsInfo } from '@prisma/client';
 import { PrismaService } from '@project-lc/prisma-orm';
 import { GoodsInfoDto } from '@project-lc/shared-types';
 import {
-  getImgSrcListFromHtmlStringList,
-  getS3KeyListFromImgSrcList,
   S3Service,
 } from '@project-lc/nest-modules-s3';
 import { ServiceBaseWithCache } from '@project-lc/nest-core';
@@ -107,7 +105,7 @@ export class GoodsCommonInfoService extends ServiceBaseWithCache {
       const imgSrcList: string[] = getImgSrcListFromHtmlStringList(contentList);
 
       // img src에서 s3에 저장된 이미지만 찾기
-      const s3ImageKeys = getS3KeyListFromImgSrcList(imgSrcList);
+      const s3ImageKeys = this.s3service.getGoodsImageS3KeyListFromImgSrcList(imgSrcList);
 
       // s3저장된 이미지 있는경우
       if (s3ImageKeys.length > 0) {
@@ -130,3 +128,7 @@ export class GoodsCommonInfoService extends ServiceBaseWithCache {
     }
   }
 }
+function getImgSrcListFromHtmlStringList(contentList: string[]): string[] {
+  throw new Error('Function not implemented.');
+}
+
