+++ libs/nest-modules-goods/src/goods/goods.service.ts
@@ -24,8 +24,6 @@ import {
   TotalStockInfo,
 } from '@project-lc/shared-types';
 import {
-  getImgSrcListFromHtmlStringList,
-  getS3KeyListFromImgSrcList,
   S3Service,
 } from '@project-lc/nest-modules-s3';
 import { ServiceBaseWithCache } from '@project-lc/nest-core';
@@ -351,7 +349,7 @@ export class GoodsService extends ServiceBaseWithCache {
     const imgSrcList: string[] = getImgSrcListFromHtmlStringList(contentList);
 
     // img src에서 s3에 저장된 이미지만 찾기
-    const s3ImageKeys = getS3KeyListFromImgSrcList(imgSrcList);
+    const s3ImageKeys = this.s3service.getGoodsImageS3KeyListFromImgSrcList(imgSrcList);
 
     if (s3ImageKeys.length > 0) {
       await this.s3service.deleteMultipleObjects(
@@ -377,7 +375,7 @@ export class GoodsService extends ServiceBaseWithCache {
     const imageList = images.map(({ image }) => image);
 
     // 이미지 중 s3에 업로드된 이미지 찾기
-    const s3ImageKeys = getS3KeyListFromImgSrcList(imageList);
+    const s3ImageKeys = this.s3service.getGoodsImageS3KeyListFromImgSrcList(imageList);
 
     if (s3ImageKeys.length > 0) {
       await this.s3service.deleteMultipleObjects(
@@ -691,3 +689,7 @@ export class GoodsService extends ServiceBaseWithCache {
     });
   }
 }
+function getImgSrcListFromHtmlStringList(contentList: string[]): string[] {
+  throw new Error('Function not implemented.');
+}
+
